+++ pkg/platform/buff/convert.go
@@ -40,14 +40,15 @@ func toSellTrade(item sellOrderItem, goodsInfos map[string]goodInfo) platform.Tr
 	}
 	price, _ := strconv.ParseFloat(item.Price, 64)
 	fee, _ := strconv.ParseFloat(item.Fee, 64)
-	income, _ := strconv.ParseFloat(item.Income, 64)
+	// TODO: is_p2p
+	// income, _ := strconv.ParseFloat(item.Income, 64)
 	return platform.TradeRecord{
 		CS2Item:      toCS2Item(item.AssetInfo, item.GoodsID, goods),
 		ExternalID:   fmt.Sprintf("buff-sell-%s", item.ID),
 		TradeType:    model.DirectionSell,
 		Quantity:     1,
 		UnitPrice:    int64(price * 100),
-		TotalPrice:   int64(income * 100),
+		TotalPrice:   int64(price * 100),
 		Fee:          int64(fee * 100),
 		TradeAt:      item.TransactTime * 1000,
 		State:        item.State,